refactor(web): move shell-outs behind infra adapters (#535 follow-up 3)
- services/git.ts shim deleted. All callers (branch-status-poller,
  sync-state, github-graphql, workspace-service, system router, the two
  test files) now import from infra/git/git-client directly.
- services/setup-runner.ts moved to infra/setup/setup-runner.ts (it is
  a bash process pool — pure infra). project-config.ts moved alongside.
- services/system-service.ts split: the raw execFile shell-outs (du,
  brew install) moved into infra/process/{du,install}.ts.
  SystemService keeps the orchestration (du semaphore, checkPrereqs)
  and delegates to infra.

Refs: https://github.com/band-app/band/issues/535
