+++ apps/web/src/server/infra/setup/project-config.ts
