+++ apps/web/src/server/services/branch-status-poller.ts
@@ -3,7 +3,7 @@ import { eq } from "drizzle-orm";
 import { toWorkspaceId } from "@/dashboard";
 import { getDb } from "../infra/db/connection";
 import { branchStatuses as branchStatusesTable } from "../infra/db/schema";
-import { execGh, execGit, getRepoInfo, type RepoInfo } from "./git";
+import { execGh, execGit, getRepoInfo, type RepoInfo } from "../infra/git/git-client";
 import { buildBatchedCIQuery, type CIStatus, parseBatchedCIResponse } from "./github-graphql";
 import { loadState } from "./state";
 import { syncWorktrees } from "./sync-state";
