+++ apps/web/src/server/services/watcher.ts
@@ -8,7 +8,7 @@ import {
   subscribe as subscribeRaw,
 } from "../infra/events/status-event-bus";
 import { startBranchStatusPoller, stopBranchStatusPoller } from "./branch-status-poller";
-import { getRunningSetups } from "./setup-runner";
+import { getRunningSetups } from "../infra/setup/setup-runner";
 import { loadCurrentStatuses } from "./state";
 
 /**
